requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
autopep8==1.6.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
Markdown==3.3.4
msgpack==1.0.3
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygame==2.0.1
pyparsing==3.0.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
